// src/Models/cafe.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0' at 'Thu, 07 Apr 2022 08:01:06 GMT'
 */
export type Cafe = IContentItem<{
  phone: Elements.TextElement;
  city: Elements.TextElement;
  photo: Elements.AssetsElement;
  email: Elements.TextElement;
  country: Elements.TextElement;
  street: Elements.TextElement;
  state: Elements.TextElement;
  zipCode: Elements.TextElement;
  sitemap: Elements.TaxonomyElement;
}>;